dev: Move cspell definitions out of nested directory
