+++ packages/cspell-types/src/CSpellSettingsDef.ts
