+++ packages/cspell-types/src/index.ts
@@ -1,3 +1,3 @@
-export * from './settings';
+export * from './CSpellSettingsDef';
 export * from './CSpellReporter';
 export * from './TextOffset';
