--- packages/cspell-types/src/settings/index.ts
@@ -1,1 +1,0 @@
-export * from './CSpellSettingsDef';
